feat: Closes #238. Move storybook.requires.js and Storybook.tsx files from the root directory to the .storybook folder.
