+++ examples/native/App.tsx
@@ -1,3 +1,3 @@
-import StorybookUIRoot from './Storybook';
+import StorybookUIRoot from './.storybook/Storybook';
 
 export { StorybookUIRoot as default };
